backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-adk>=0.3.0
google-genai>=0.5.0
google-cloud-firestore>=2.14.0
pydantic>=2.5.0
python-dotenv>=1.0.0
websockets>=12.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
